firstcommit

+++ database/migrations/2025_02_22_081202_add_google2fa_to_users_table.php
@@ -4,20 +4,25 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration
-{
+return new class extends Migration {
+    public function up()
+    {
+        Schema::table('users', function (Blueprint $table) {
+            $table->boolean('google2fa_enabled')->default(false);
+            $table->string('google2fa_secret')->nullable()->after('password');
+        });
+    }
 
-   public function up()
-{
-    Schema::table('users', function (Blueprint $table) {
-        $table->string('google2fa_secret')->nullable();
-    });
-}
-public function down()
-{
-    Schema::table('users', function (Blueprint $table) {
-        $table->dropColumn('google2fa_secret');
-    });
-}
-
+    public function down()
+    {
+        Schema::table('users', function (Blueprint $table) {
+            if (Schema::hasColumn('users', 'google2fa_enabled')) {
+                $table->dropColumn('google2fa_enabled');
+            }
+            if (Schema::hasColumn('users', 'google2fa_secret')) {
+                $table->dropColumn('google2fa_secret');
+            }
+        });
+    }
+    
 };
